+++ stu-manager/src/main/java/xin/xlchen/dhu/stumanger/service/TermService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import xin.xlchen.dhu.stumanger.mapper.TermMapper;
+import xin.xlchen.dhu.stumanger.model.MCourse;
+import xin.xlchen.dhu.stumanger.model.MRelTermCourse;
 import xin.xlchen.dhu.stumanger.model.MResult;
 import xin.xlchen.dhu.stumanger.model.MTerm;
 
@@ -41,10 +43,16 @@ public class TermService {
     public MResult deleteTerm(String termId){
     	MResult mResult = new MResult();
     	try {
-        	//删除数据
-        	termMapper.deleteTerm(termId);
-            //
-            mResult.setSuccess(true);
+    		if (termMapper.getUsedTermCount(termId) > 0) {
+    			///学籍已在使用中,不允许删除
+    			mResult.setSuccess(false);
+    			mResult.setErrorMsg("学籍已在使用中,不允许删除!");
+    		} else {
+	        	//删除数据
+	        	termMapper.deleteTerm(termId);
+	            //
+	            mResult.setSuccess(true);
+    		}
 		} catch (Exception e) {
 			mResult.setSuccess(false);
 			mResult.setErrorMsg(e.getMessage());
@@ -92,6 +100,53 @@ public class TermService {
     			//编辑
     			termMapper.editTerm(term);
     		}
+		} catch (Exception e) {
+			mResult.setSuccess(false);
+			mResult.setErrorMsg(e.getMessage());
+		}
+        return mResult;
+    }
+    
+    /**
+     * 查询指定学期已分配的课程列表
+     * @return
+     */
+    public List<MCourse> getCourseInTerm(String termId){
+    	return termMapper.getCourseInTerm(termId);
+    }
+    
+    /**
+     * 查询指定学期未分配的课程列表
+     * @return
+     */
+    public List<MCourse> getCourseNotInTerm(String termId){
+    	return termMapper.getCourseNotInTerm(termId);
+    }
+    
+    
+    /**
+     * 学籍添加课程
+     * @param mRelTermCourse
+     */
+    public MResult addTermCourse(MRelTermCourse mRelTermCourse) {
+     	MResult mResult = new MResult();
+    	try {
+    		termMapper.addTermCourse(mRelTermCourse);
+		} catch (Exception e) {
+			mResult.setSuccess(false);
+			mResult.setErrorMsg(e.getMessage());
+		}
+        return mResult;
+    }
+    
+    /**
+     * 学籍删除课程
+     * @param mRelTermCourse
+     */
+    public MResult removeTermCourse(MRelTermCourse mRelTermCourse) {
+      	MResult mResult = new MResult();
+    	try {
+    		termMapper.removeTermCourse(mRelTermCourse);
 		} catch (Exception e) {
 			mResult.setSuccess(false);
 			mResult.setErrorMsg(e.getMessage());
